+++ CHANGELOG.md
@@ -4,8 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- ACF Pro can now be updated via `composer` as well as the legacy method
 - Added `--skip-git` for projects that are focused only on client communication
 - Added ability to turn off RSS feed per project, even if set globally
+
 ### [Fixed]
 - Fixed a bug in preparing new Wordpress projects
 - Fixed an activity log email formatting error